+++ itests/app_two.py
@@ -17,7 +17,7 @@ from blacksheep.server.bindings import (
     FromCookie,
     FromForm,
     FromHeader,
-    FromJson,
+    FromJSON,
     FromQuery,
 )
 from blacksheep.server.controllers import ApiController, delete, get, post
@@ -30,6 +30,7 @@ from blacksheep.server.openapi.common import (
     ResponseInfo,
 )
 from blacksheep.server.openapi.v3 import OpenAPIHandler
+from blacksheep.server.openapi.ui import ReDocUIProvider
 from dateutil.parser import parse as dateutil_parse
 from guardpost.authentication import Identity
 from guardpost.authorization import AuthorizationContext
@@ -51,6 +52,7 @@ app_two = Application()
 
 # OpenAPI v3 configuration:
 docs = OpenAPIHandler(info=Info(title="Cats API", version="0.0.1"))
+docs.ui_providers.append(ReDocUIProvider())
 
 # include only endpoints whose path starts with "/api/"
 docs.include = lambda path, _: path.startswith("/api/")
@@ -298,6 +300,19 @@ class CatPetModel(PetModel):
     laziness: float
 
 
+@dataclass
+class Foo:
+    id: UUID
+    name: str
+    cool: float
+
+
+@dataclass
+class FooList:
+    items: List[Foo]
+    total: int
+
+
 def on_polymorph_example_docs_created(
     docs: OpenAPIHandler, operation: Operation
 ) -> None:
@@ -386,6 +401,40 @@ class Cats(ApiController):
         Returns a list of paginated cats.
         """
 
+    @get("foos")
+    def get_foos() -> FooList:
+        ...
+
+    @get("cats2")
+    def get_cats_alt(
+        self,
+        page: FromQuery[int] = FromQuery(1),
+        page_size: FromQuery[int] = FromQuery(30),
+        search: FromQuery[str] = FromQuery(""),
+    ) -> CatsList:
+        """
+        Alternative way to have the response type for status 200 documented.
+        """
+        return CatsList(
+            [
+                Cat(
+                    id=UUID("3fa85f64-5717-4562-b3fc-2c963f66afa6"),
+                    name="Foo",
+                    active=True,
+                    type=CatType.EUROPEAN,
+                    creation_time=dateutil_parse("2020-10-25T19:39:31.751652"),
+                ),
+                Cat(
+                    id=UUID("f212cabf-987c-48e6-8cad-71d1c041209a"),
+                    name="Frufru",
+                    active=True,
+                    type=CatType.PERSIAN,
+                    creation_time=dateutil_parse("2020-10-25T19:39:31.751652"),
+                ),
+            ],
+            1230,
+        )
+
     @docs.ignore()
     @get("/ignored")
     def secret_api(self):
@@ -437,7 +486,7 @@ class Cats(ApiController):
 
     @post()
     @docs(create_cat_docs)
-    def create_cat(self, input: FromJson[CreateCatInput]) -> Response:
+    def create_cat(self, input: FromJSON[CreateCatInput]) -> Response:
         """
         Creates a new cat.
         """